ref(node): Remove backend
