--- packages/node/src/backend.ts
@@ -1,58 +1,0 @@
-import { BaseBackend } from '@sentry/core';
-import { Event, EventHint, SeverityLevel, Transport, TransportOptions } from '@sentry/types';
-import { Dsn } from '@sentry/utils';
-
-import { eventFromException, eventFromMessage } from './eventbuilder';
-import { HTTPSTransport, HTTPTransport } from './transports';
-import { NodeOptions } from './types';
-
-/**
- * The Sentry Node SDK Backend.
- * @hidden
- */
-export class NodeBackend extends BaseBackend<NodeOptions> {
-  /**
-   * @inheritDoc
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/explicit-module-boundary-types
-  public eventFromException(exception: any, hint?: EventHint): PromiseLike<Event> {
-    return eventFromException(this._options, exception, hint);
-  }
-
-  /**
-   * @inheritDoc
-   */
-  public eventFromMessage(message: string, level: SeverityLevel = 'info', hint?: EventHint): PromiseLike<Event> {
-    return eventFromMessage(this._options, message, level, hint);
-  }
-
-  /**
-   * @inheritDoc
-   */
-  protected _setupTransport(): Transport {
-    if (!this._options.dsn) {
-      // We return the noop transport here in case there is no Dsn.
-      return super._setupTransport();
-    }
-
-    const dsn = new Dsn(this._options.dsn);
-
-    const transportOptions: TransportOptions = {
-      ...this._options.transportOptions,
-      ...(this._options.httpProxy && { httpProxy: this._options.httpProxy }),
-      ...(this._options.httpsProxy && { httpsProxy: this._options.httpsProxy }),
-      ...(this._options.caCerts && { caCerts: this._options.caCerts }),
-      dsn: this._options.dsn,
-      tunnel: this._options.tunnel,
-      _metadata: this._options._metadata,
-    };
-
-    if (this._options.transport) {
-      return new this._options.transport(transportOptions);
-    }
-    if (dsn.protocol === 'http') {
-      return new HTTPTransport(transportOptions);
-    }
-    return new HTTPSTransport(transportOptions);
-  }
-}
